// client/rest/models/tags-jsonld.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Tags
 * @export
 * @interface TagsJsonld
 */
export interface TagsJsonld {
    /**
     * 
     * @type {string | any}
     * @memberof TagsJsonld
     */
    '@context'?: string | any;
    /**
     * 
     * @type {string}
     * @memberof TagsJsonld
     */
    '@id'?: string;
    /**
     * 
     * @type {string}
     * @memberof TagsJsonld
     */
    '@type'?: string;
    /**
     * 
     * @type {number}
     * @memberof TagsJsonld
     */
    'tagId'?: number;
    /**
     * 
     * @type {string}
     * @memberof TagsJsonld
     */
    'name'?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof TagsJsonld
     */
    'recipe'?: Array<string>;
}