// client/rest/models/images-jsonld.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ImagesJsonld
 */
export interface ImagesJsonld {
    /**
     * 
     * @type {string | any}
     * @memberof ImagesJsonld
     */
    '@context'?: string | any;
    /**
     * 
     * @type {string}
     * @memberof ImagesJsonld
     */
    '@id'?: string;
    /**
     * 
     * @type {string}
     * @memberof ImagesJsonld
     */
    '@type'?: string;
    /**
     * 
     * @type {number}
     * @memberof ImagesJsonld
     */
    'imageId'?: number;
    /**
     * 
     * @type {string}
     * @memberof ImagesJsonld
     */
    'date'?: string;
    /**
     * 
     * @type {string}
     * @memberof ImagesJsonld
     */
    'path'?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof ImagesJsonld
     */
    'recipe'?: Array<string>;
}