// client/rest/models/ingredients-jsonld.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Ingredients
 * @export
 * @interface IngredientsJsonld
 */
export interface IngredientsJsonld {
    /**
     * 
     * @type {string | any}
     * @memberof IngredientsJsonld
     */
    '@context'?: string | any;
    /**
     * 
     * @type {string}
     * @memberof IngredientsJsonld
     */
    '@id'?: string;
    /**
     * 
     * @type {string}
     * @memberof IngredientsJsonld
     */
    '@type'?: string;
    /**
     * 
     * @type {number}
     * @memberof IngredientsJsonld
     */
    'ingredientId'?: number;
    /**
     * 
     * @type {string}
     * @memberof IngredientsJsonld
     */
    'name'?: string;
}